+++ fetch/fetch.go
@@ -1,7 +1,10 @@
 package fetch
 
 import (
+	"fmt"
 	"time"
+	"runtime"
+	"log"
 
 	"github.com/briandowns/spinner"
 	"github.com/secman-team/shell"
@@ -12,7 +15,19 @@ func FetchSECDIR() {
 	s.Suffix = " Fetching..."
 	s.Start()
 
-	shell.SHCore("secman-sync pl", "bash $HOME/sm/ssc pl")
-
+	err, out, errout := shell.ShellOut("")
+	
+	if runtime.GOOS == "windows" {
+		err, out, errout = shell.PWSLOut("~/sm/secman-sync.ps1 --sm")
+	} else {
+		err, out, errout = shell.ShellOut("secman-sync pl")
+	}
+		
+	if err != nil {
+		log.Printf("error: %v\n", err)
+		fmt.Println(errout)
+	}
+		
 	s.Stop()
+	fmt.Println(out)
 }