+++ upgrade/upg.go
@@ -1,7 +1,9 @@
 package upg
 
-import "github.com/secman-team/shell"
+import (
+	"github.com/secman-team/shell"
+)
 
 func Upgrade() {
-	shell.SHCore("verx --upg", "bash $HOME/sm/vx --upg")
+	shell.ShellCmd("verx --sm")
 }